dialer/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
phonenumbers==8.13.36
pydantic==1.10.14
jinja2==3.1.3
httpx==0.27.0
twilio==9.0.4
prompt_toolkit==3.0.43
sqlite-utils==3.36
